refactor name on the UI so its consistant with the other naming Management UI's
