+++ UI_Layer/SupervisorUI.py
@@ -3,7 +3,7 @@ import os
 import shutil
 import sys
 
-from UI_Layer.ContractorUI import ContractorUI
+from UI_Layer.ContractorManagment_UI import ContractorUI
 from UI_Layer.EmployeeManagement_UI import EmployeeManagementUI
 from UI_Layer.MaintenanceReportManagement_UI import MaintenanceReportUI
 from UI_Layer.PropertyManagement_UI import PropertyUI
